Stac 12220 - make state location configurable from check config (#123)
* make state location configurable from check config

* add state to agent-integration-sample

* persistent state tests

* retry test

* fix windows error message

* drop test, not working with sudo on linux container

* only try to remove file if it exists

+++ stackstate_checks_base/stackstate_checks/base/utils/persistent_state.py
@@ -56,19 +56,15 @@ class StateManager:
         if instance.instance_key in self.state:
             del self.state[instance.instance_key]
 
-        try:
-            os.remove(instance.file_location)
-            self.log.debug("PersistentState: Removed state for instance: {}".format(instance.instance_key))
-        except OSError as e:
-            # File not found / no file for this state so the state doesn't exist. Catch exception and return None
-            if e.errno == errno.ENOENT:
-                self.log.debug("PersistentState: No state file found for instance: {} expecting it at: {}. {}"
-                               .format(instance.instance_key, instance.file_location, e))
-                return None
-
-            self.log.error("PersistentState: Failed to remove state file for instance: {}. {}"
-                           .format(instance.instance_key, e))
-            raise StateClearException(e)
+        # if the file exists, try to delete it.
+        if os.path.isfile(instance.file_location):
+            try:
+                os.remove(instance.file_location)
+                self.log.debug("PersistentState: Removed state for instance: {}".format(instance.instance_key))
+            except OSError as e:
+                self.log.error("PersistentState: Failed to remove state file for instance: {}. {}"
+                               .format(instance.instance_key, e))
+                raise StateClearException(e)
 
     def _read_state(self, instance):
         """
@@ -83,7 +79,7 @@ class StateManager:
             self.log.error("PersistentState: State file is corrupted for instance: {} stored at: {}. {}"
                            .format(instance.instance_key, instance.file_location, e))
             raise StateCorruptedException(e)
-        except IOError as e:
+        except (OSError, IOError) as e:
             # File not found / no file for this state so the state doesn't exist. Catch exception and return None
             if e.errno == errno.ENOENT or e.errno == errno.EINVAL:
                 self.log.debug("PersistentState: No state file found for instance: {}"
@@ -150,7 +146,7 @@ class StateManager:
 
                 with open(instance.file_location, 'w+') as f:
                     f.write(json.dumps(self.state[instance.instance_key]))
-            except OSError as e:
+            except (IOError, OSError) as e:
                 if e.errno != errno.EEXIST:
                     # if we couldn't save, log the state
                     self.log.error('PersistentState: Could not persist state for instance: {} at {}, '
